requirements.txt
# ===== Web Framework =====
Django==5.1.2
django-ninja==1.5.0
gunicorn==21.2.0
django-environ==0.11.2
django-storages==1.14.2
boto3==1.34.34

# ===== Whisper STT =====
faster-whisper==1.0.0
transformers==4.36.0

# ===== Audio Processing =====
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
audioread==3.0.1

# ===== Utilities =====
numpy==1.26.4
requests==2.32.3
PyYAML==6.0.2
python-dateutil==2.9.0

# Note: PyTorch는 Dockerfile에서 별도로 CPU 버전 설치
